Fix case where state != enabled caused error
When using the register_site state, if the `state` for the
site in the apache pillar did not equal "enabled" the error

```
Cannot extend ID 'a2ensite <name>' in 'base:apache.register_site'. It is not part of the high state.
```

If statements were there to build out the `a2ensite` or `a2dissite` ID
appropiately, but the next state for managing the config always
assumed that the `a2ensite` id would be there.

Addressed by making a Jinja variable that will be the appropiate ID,
probably could use a better name.