Include the mod_* states in apache.modules for Debian hosts
The formula provides an option to enable Apache modules using pillar
data, but this does not install packages dependencies as well. This can
result is states failing to be applied because packages are missing.

By including the mod_* states in the apache.modules the package
dependencies should be handled correctly and it improves DRY of a2enmod
cmds.

This is only applied for Debian hosts because the individual mod_*
states don't seem to support enabling modules for RedHat hosts.